deploy: add lambda deploy type
Add a new form of Deploy: Lambda. It performs the following actions:

1. Take an array of files and folders (deploy.package) and zips it.
2. Sends the zip to AWS S3 to a specified bucket (deploy.s3_bucket) and to
an S3 key that is the sha of the commit. E.g. when deploying commit abcd123
to the bucket my-bucket, the deploy will create a key on AWS S3 named
my-bucket/abcd123/package.zip. The sha of the commit comes from the SCM step.
3. Take an array of lambda functions (deploy.functions) and build the name of
the functions with the template ${environment}-${projectName}-${functionName}.
E.g. given the function name hello being deployed on the acme project on the
live environment, the final name of the function will be live-acme-hello.
4. Updates the functions code on AWS Lambda.